+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/tests/latest/test_storage_account_scenarios.py
@@ -285,3 +285,22 @@ class StorageAccountTests(StorageScenarioMixin, ScenarioTest):
         with self.assertRaises(SystemExit) as ex:
             self.cmd('storage account show -g {rg} -n fake_account')
         self.assertEqual(ex.exception.code, 3)
+
+    @ResourceGroupPreparer()
+    def test_management_policy(self, resource_group):
+        import os
+        from msrestazure.azure_exceptions import CloudError
+        curr_dir = os.path.dirname(os.path.realpath(__file__))
+        policy_file = os.path.join(curr_dir, 'mgmt_policy.json').replace('\\', '\\\\')
+
+        storage_account = self.create_random_name(prefix='cli', length=24)
+
+        self.kwargs = {'rg': resource_group, 'sa': storage_account, 'policy': policy_file}
+        self.cmd('storage account create -g {rg} -n {sa} --kind StorageV2')
+        self.cmd('storage account management-policy create --account-name {sa} -g {rg} --policy @"{policy}"')
+        self.cmd('storage account management-policy update --account-name {sa} -g {rg}'
+                 ' --set "policy.rules[0].name=newname"')
+        self.cmd('storage account management-policy show --account-name {sa} -g {rg}',
+                 checks=JMESPathCheck('policy.rules[0].name', 'newname'))
+        self.cmd('storage account management-policy delete --account-name {sa} -g {rg}')
+        self.cmd('storage account management-policy show --account-name {sa} -g {rg}', expect_failure=True)