+++ src/command_modules/azure-cli-servicefabric/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.1.19
+++++++
+* Upgrade azure-mgmt-storage from 3.1.1 to 3.3.0
+
 0.1.18
 ++++++
 * Minor fixes.